feat(threads): build the Unchanged block once, carry it on the context
generateAll fans one ActivityContext across every style, so the block is
built alongside the dossier and merely emitted per voice later. Building it
per voice would quietly undo the single-pass work from PR #65.

ThreadsDossierBuilder.build keeps its old signature (all ~12 existing
callers untouched); the dual dossier+unchangedBlock computation lives in a
new buildResult, which PromptListView's one production call site now reads
from instead of building twice. The memo carries unchangedBlock through
cache hits too, so a same-session reopen doesn't silently drop it.

The pre-existing DEBUG field-report harness moved to its own file
(ThreadsDossierFieldReport.swift) to keep the builder under the file_length
gate after the dual-return plumbing landed — pure move, no behavior change.
